Implemented several asic commands.

+++ src/main/scala/com/mikrotasarim/camera/device/DeviceInterface.scala
@@ -1,7 +1,7 @@
 package com.mikrotasarim.camera.device
 
 trait DeviceInterface {
-  def SetWireInValue(wireNumber: Int, value: Int)
+  def SetWireInValue(wireNumber: Int, value: Long)
   def ActivateTriggerIn(address: Int, bit: Int)
   def UpdateWireIns()
 }
